Included support for Updatehub library
This library is a support for the UPDATEHUB platform in zephyr. The library
has the ability to query the UPDATEHUB server for the purpose of verifying
that it has some update to the board. Having an update the UPDATEHUB library
itself is able to download the update and save it on the board, everything
happens correctly the library will restart the card and the new image will
start working.